seminars/seminar4-event-driven/queues/requirements.txt
requests==2.31.0
pika==1.3.2
minio==7.2.0
Pillow==10.0.0
